Testing done, server runs on linux
